34:fix:divided the helper functions into separate files and other refactoration
