+++ lib/core/utils/helper_functions/form_mixin.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+
+mixin FormStateMixin<T extends StatefulWidget> on State<T> {
+  final GlobalKey<FormState> formKey = GlobalKey<FormState>();
+  bool isSubmitting = false;
+
+  bool validateForm() {
+    return formKey.currentState?.validate() ?? false;
+  }
+
+  void setSubmitting(bool submitting) {
+    if (mounted) {
+      setState(() {
+        isSubmitting = submitting;
+      });
+    }
+  }
+
+  void showSuccessMessage(String message) {
+    if (mounted) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text(message),
+          backgroundColor: Colors.green,
+          behavior: SnackBarBehavior.floating,
+        ),
+      );
+    }
+  }
+
+  void showErrorMessage(String message) {
+    if (mounted) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text(message),
+          backgroundColor: Colors.red,
+          behavior: SnackBarBehavior.floating,
+        ),
+      );
+    }
+  }
+
+  void unfocusAll() {
+    FocusScope.of(context).unfocus();
+  }
+}
