+++ lib/features/resume/widgets/final_step_dialog.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:interview/app/themes/fonts.dart';
 import 'package:interview/app/themes/text_styles.dart';
 import 'package:interview/core/constants/colors.dart';
 import 'package:interview/core/utils/extensions/responsive_extension.dart';
